Split typing into two files
